requirements.txt
discord
PyYAML
regex